Revert "Making sure models aren't requested twice when cache_classes is true."
When cache_classes is true, we were not seeing any models loaded and the result was that an thinking_sphinx:configure resulted in a blank configuration file -- the pre-existing behavior, which eager-loaded AR classes, was necessary in our environment. This eager loading only applies to configuration and not to using the thinking sphinx api.

This reverts commit 501b4e23d8c8bc0d949fd5fc55a807a84a5da745.